@@ -16,24 +16,25 @@
   "private": true,
   "dependencies": {
     "@angular-devkit/core": "^0.6.8",
-    "@angular/animations": "5.2.9",
-    "@angular/common": "5.2.9",
-    "@angular/compiler": "5.2.9",
-    "@angular/core": "5.2.9",
-    "@angular/forms": "5.2.9",
-    "@angular/platform-browser": "5.2.9",
-    "@angular/platform-browser-dynamic": "5.2.9",
-    "@angular/router": "5.2.9",
+    "@angular/animations": "6.0.7",
+    "@angular/common": "6.0.7",
+    "@angular/compiler": "6.0.7",
+    "@angular/core": "6.0.7",
+    "@angular/forms": "6.0.7",
+    "@angular/platform-browser": "6.0.7",
+    "@angular/platform-browser-dynamic": "6.0.7",
+    "@angular/router": "6.0.7",
     "core-js": "^2.4.1",
     "marked": "^0.3.9",
-    "rxjs": "^5.5.8",
-    "zone.js": "^0.8.25"
+    "rxjs": "^6.2.1",
+    "zone.js": "^0.8.26",
+    "rxjs-compat": "^6.0.0-rc.0"
   },
   "devDependencies": {
     "@angular-devkit/build-angular": "~0.7.0-beta.2",
     "@angular/cli": "^6.1.0-rc.0",
-    "@angular/compiler-cli": "5.2.9",
-    "@angular/language-service": "5.2.9",
+    "@angular/compiler-cli": "6.0.7",
+    "@angular/language-service": "6.0.7",
     "@types/jasmine": "~2.5.53",
     "@types/jasminewd2": "~2.0.2",
     "@types/node": "^9.4.0",
@@ -50,6 +51,6 @@
     "protractor": "~5.1.2",
     "ts-node": "~4.1.0",
     "tslint": "~5.9.1",
-    "typescript": "2.6.2"
+    "typescript": "2.7.2"
   }
 }
