@@ -1,9 +1,9 @@
 import { Injectable } from '@angular/core';
-import { Observable } from 'rxjs/Observable';
+import { Observable } from 'rxjs';
 
 import { ApiService } from './api.service';
 import { Comment } from '../models';
-import { map } from 'rxjs/operators/map';
+import { map } from 'rxjs/operators';
 
 
 @Injectable()
